Factor out validation logic to separate file
